Working and tested Knights Tour code
This code is capable of completing the knights tour problem, though it does take a good while to run and process the outputs.

This method used iteration to solve the problem. While researching backtracking, recursion was mentioned. I may convert this into a recursive solution to see if it improves the speed
